+++ docs/user/metal/install_upi.md
@@ -202,10 +202,10 @@ TODO RHEL CoreOS does not have assets for bare-metal.
 
 ### Monitor for bootstrap-complete
 
-The administrators can use the `upi bootstrap-complete` target of the OpenShift Installer to monitor cluster bootstrapping. The command succeeds when it notices `bootstrap-complete` event from Kubernetes APIServer. This event is generated by the bootstrap machine after the Kubernetes APIServer has been bootstrapped on the control plane machines. For example,
+The administrators can use the `wait-for bootstrap-complete` target of the OpenShift Installer to monitor cluster bootstrapping. The command succeeds when it notices `bootstrap-complete` event from Kubernetes APIServer. This event is generated by the bootstrap machine after the Kubernetes APIServer has been bootstrapped on the control plane machines. For example,
 
 ```console
-$ openshift-install --dir test-bare-metal upi bootstrap-complete
+$ openshift-install --dir test-bare-metal wait-for bootstrap-complete
 INFO Waiting up to 30m0s for the Kubernetes API at https://api.test.example.com:6443...
 INFO API v1.12.4+c53f462 up
 INFO Waiting up to 30m0s for the bootstrap-complete event...
@@ -213,10 +213,10 @@ INFO Waiting up to 30m0s for the bootstrap-complete event...
 
 ## Monitor for cluster completion
 
-The administrators can use the `upi finish` target of the OpenShift Installer to monitor cluster completion. The command succeeds when it notices that Cluster Version Operator has completed rolling out the OpenShift cluster from Kubernetes APIServer.
+The administrators can use the `wait-for cluster-ready` target of the OpenShift Installer to monitor cluster completion. The command succeeds when it notices that Cluster Version Operator has completed rolling out the OpenShift cluster from Kubernetes APIServer.
 
 ```console
-$ openshift-install upi finish
+$ openshift-install wait-for cluster-ready
 INFO Waiting up to 30m0s for the cluster to initialize...
 ```
